Add files via upload
This is the first step of my final master project in order to find experts in different field (exp, java).
first step is learn how to work with Lucene.
I have a data-set with ~32G of stack-overflow from 2008 to 2015 .
In this step the goal is to find answer for tester questions.
step 1. parse Posts.xml by FinalXMLParser , it gives us a txt file as output. 
step 2. index txt files
step 3. run Tester to see the result.